requirements.txt
streamlit
langchain
langgraph
openai
chromadb
langchain-community
pydantic
python-dotenv
pandas
numpy
langchain-openai